[cross] Grind main/gbaque: 95.36% -> 95.85% (#16951)
* Match MakeSellData closer: writePtr copy, signedness of execParam/price compares


* Match MakeBuyData closer: foodBasePtr re-reads, work/packed swap idiom, inline TableStrings


* Improve MakeSmithData k-loop match via explicit unroll


* Split lbl_801DB418 into named gbaque error strings in symbols.txt


* [cross] GetItemIcon returns int (matches caller codegen in gbaque and menu_cmd)


* gbaque: claim GBA rodata block, OpenMenu via GbaQue, GetPlayerHP/LoadPlayerStat fixes, verified permute wins

- Claim orphaned 0x801DB2E0-0x801DB370 rodata (GBA path strings + letter item
  tables) for gbaque.cpp; split lbl_801DB2E0 in symbols.txt, extend splits.txt
- OpenMenu: semaphore/m_singleMode accessed through GbaQue global (88.99->96.43)
- GetPlayerHP: duplicated hp!=prevHp check and u8-cast return (90.42->98.13)
- LoadPlayerStat: progress clamp compares <=0xFF (cmplwi 0xff/bgt)
- Verified permute candidates kept (LoadAll, ExecutQueue, GetPauseMode,
  GetControllerMode, LoadPlayerStat l0/l6); semantic flips rejected
  (bossStageLimit cmplw, GetScouterInfo lha, execParam cmpwi x2)


---------

+++ src/singmenu.cpp
@@ -3563,7 +3563,7 @@ inline void CMenuPcs::SingLifeResetWait()
  * JP Address: TODO
  * JP Size: TODO
  */
-u8 CMenuPcs::GetItemIcon(int index)
+int CMenuPcs::GetItemIcon(int index)
 {
     return gSingMenuItemIconByType[index];
 }
